Resolve automated code review comments
